Remove callable classes functions.
