+++ storage_bucket/delete.py
@@ -1,28 +1,4 @@
-from attr import dataclass
-from typing_extensions import final
-
-from storage_bucket.get import GetBucket
-
-
-@final
-@dataclass(frozen=True, slots=True)
-class DeleteBucket(object):
-    """Delete a gcp storage bucket."""
-
-    get_bucket = GetBucket()
-
-    def __call__(
-        self,
-        *,
-        storage_bucket_name: str,
-        force: bool = False,
-        **kwargs,
-    ) -> None:
-        """Delete bucket."""
-        bucket = self.get_bucket(
-            storage_bucket_name=storage_bucket_name,
-        )
-        bucket.delete(force=force, **kwargs)
+from storage_bucket.bucket import get_bucket
 
 
 def delete_bucket(
@@ -31,9 +7,8 @@ def delete_bucket(
     force: bool = False,
     **kwargs,
 ) -> None:
-    """Run DeleteBucket."""
-    return DeleteBucket()(
+    """Delete a storage bucket."""
+    bucket = get_bucket(
         storage_bucket_name=storage_bucket_name,
-        force=force,
-        **kwargs,
     )
+    bucket.delete(force=force, **kwargs)
