+++ storage_bucket/delete_file.py
@@ -1,47 +1,4 @@
-from attr import dataclass
-from typing_extensions import final
-
-from storage_bucket.get import GetBucket
-
-
-@final
-@dataclass(frozen=True, slots=True)
-class DeleteFile(object):
-    """
-    Deletes a google cloud storage bucket file.
-
-    .. versionadded:: 0.2.0
-
-    :param bucket_name: Name of the google cloud bucket
-    :type bucket_name: str
-
-    :param filename: The name of the file to delete
-    :type filename: str
-
-    :param generation: ID of the generation(version) of the file to delete
-    :type generation: int
-
-    :param timout: The amount of time, in seconds, to wait for the server
-    response.
-
-    :return: None
-    """
-
-    get_bucket = GetBucket()
-
-    def __call__(
-        self,
-        storage_bucket_name: str,
-        filename: str,
-        **kwargs,
-    ) -> None:
-        """Delete storage bucket file."""
-        bucket = self.get_bucket(storage_bucket_name=storage_bucket_name)
-
-        bucket.delete_blob(
-            filename,
-            **kwargs,
-        )
+from storage_bucket.bucket import get_bucket
 
 
 def delete_file(
@@ -49,9 +6,10 @@ def delete_file(
     filename: str,
     **kwargs,
 ) -> None:
-    """Run DeleteFile."""
-    DeleteFile()(
-        storage_bucket_name=storage_bucket_name,
-        filename=filename,
+    """Delete a file."""
+    bucket = get_bucket(storage_bucket_name=storage_bucket_name)
+
+    bucket.delete_blob(
+        filename,
         **kwargs,
     )
